requirements.txt
Flask
flask-cors
python-dotenv
Flask-SQLAlchemy
SQLAlchemy
dash
pandas
plotly-express
dash-bootstrap-components
mysql-connector-python